MusChart (PC 0.3)

+++ muschart/src/main/java/by/gsu/bean/Artist.java
@@ -26,7 +26,7 @@ public class Artist extends Model {
 
     @Column(name = "rating", nullable = false, length = 5)
     @DefaultValue(value = "0")
-    private int               rating;
+    private long              rating;
 
     @ManyToMany(fetch = FetchType.LAZY, cascade = {CascadeType.DETACH, CascadeType.MERGE,
             CascadeType.REFRESH, CascadeType.PERSIST}, targetEntity = Track.class)
@@ -70,11 +70,11 @@ public class Artist extends Model {
         this.photo = photo;
     }
 
-    public int getRating() {
+    public long getRating() {
         return rating;
     }
 
-    public void setRating(final int rating) {
+    public void setRating(final long rating) {
         this.rating = rating;
     }
